--- HotlesTask/app/src/main/java/com/example/georgesamuel/hotlestask/MainActivity.java
@@ -1,14 +1,0 @@
-package com.example.georgesamuel.hotlestask;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
