+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Jul 28 12:49:45 EET 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.1.1-all.zip
